testing the api endpoints

+++ migrations/001-create-waste-collector-table.sql
@@ -1,5 +1,5 @@
 create table waste_collector(
-    id  integer primary key,
+    id  serial primary key,
     first_name text,
     last_name text,
     cell_number text,
@@ -8,7 +8,7 @@ create table waste_collector(
     location text,
     age integer,
     gender text,
-    id_number integer,
+    id_number BIGINT,
     user_password text,
     verification boolean
 );
